Add a new report for open changes by "newbie committers" only
We need an easy way to spot newcomers before it's too late
and they've lost all will to contribute because their patches
are rotting, drowned in the mass of CR backlog.

This report is just a filtered monkey-copy of the existing one.
(I'm unable to understand the python format and SQL magic
needed to just add a sortable column for merged changes.)